4/10 -

+++ GetObject.cpp
@@ -33,7 +33,7 @@ public:
     Object();
     void getAverage();
     bool getObject(int duration, char& characterList, Mat& matList);
-    double getDiff(Mat&, int&, int&,int*);
+    int getDiff(Mat&, int&, int&,int*);
 };
 
 Object::Object() {
@@ -222,7 +222,7 @@ bool Object::getObject(int duration, char& characterList, Mat& matList) {
     return false;
 }
 
-double Object::getDiff(Mat &video, int &x, int &y, int* background) { //As described above, this is a literal distance formula, given r, g, and b are used as base references
+int Object::getDiff(Mat &video, int &x, int &y, int* background) { //As described above, this is a literal distance formula, given r, g, and b are used as base references
     
     Vec3b tmp_color = video.at<cv::Vec3b>(Point(x,y));
     //int tmp_diff = abs(background[2] - tmp_color[0]);
@@ -231,7 +231,7 @@ double Object::getDiff(Mat &video, int &x, int &y, int* background) { //As descr
     //std::cout << "O:" << (int)tmp_color[0] << std::endl;
     //int tmp_total = tmp_color[0] + tmp_color[1] + tmp_color[2] + 1;
     //double scaling_factor = 3*255/tmp_total;
-    double greyness = abs(tmp_color[0]-tmp_color[2])+abs(tmp_color[2]-tmp_color[1])+abs(tmp_color[1]-tmp_color[0]);
+    int greyness = abs(tmp_color[0]-tmp_color[2])+abs(tmp_color[2]-tmp_color[1])+abs(tmp_color[1]-tmp_color[0]);
     //abs(255 - tmp_color[0]*scaling_factor) + abs(255 - tmp_color[1]*scaling_factor) + abs(255 - tmp_color[2]*scaling_factor);
     std::cout << "O:" << greyness << std::endl;
     return greyness;
